// src/app/models/interfaces.ts
export interface Planeta {
    name: string;
    residents: string[];
}

export interface Filme {
    title: string;
    episode_id: number;
    release_date: string;
}

export interface Personagem {
    name: string;
    height: string;
    mass: string;
    birth_year: string;
    gender: string;
    films: string[];
}